requirements.txt
appnope==0.1.4
asgiref==3.4.1
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2021.5.30
cffi==1.11.5
chardet==3.0.4
click==8.0.4
cloudpickle==0.8.0
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.6
dask==1.1.4
dataclasses==0.8
decorator==4.4.0
defusedxml==0.5.0
easyocr==1.3
entrypoints==0.3
fastapi==0.68.2
graphviz==0.8.4
h11==0.13.0
h5py==2.8.0
h5py-cache==1.0
icc-rt==2019.0
idna==2.7
importlib-metadata==4.8.3
intel-numpy==1.14.3
intel-openmp==2023.2.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
lmdb==0.94
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.6
numpy==1.14.5
olefile==0.46
opencv-python==3.4.2.17
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
ptyprocess==0.6.0
pycparser==2.19
pydantic==1.9.2
Pygments==2.3.1
pyOpenSSL==18.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
PySocks==1.6.8
python-bidi==0.4.2
python-dateutil==2.8.0
python-multipart==0.0.5
pytz==2018.9
PyWavelets==1.0.2
PyYAML==6.0.1
pyzmq==17.1.2
requests==2.20.1
scikit-image==0.14.2
scipy==1.1.0
Send2Trash==1.5.0
Shapely==1.8.5.post1
six==1.11.0
starlette==0.14.2
svgpathtools==1.3.3
svgwrite==1.2.1
tbb==2021.10.0
tensorboardX==1.6
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==1.10.1
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.11.2
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
typing_extensions==4.1.1
urllib3==1.24.1
uvicorn==0.15.0
visdom==0.1.8.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
zipp==3.6.0
